Fix warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
visp/modules/vision/src/key-point/vpKeyPoint.cpp:812:35: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  812 |   pose.setNbParallelRansacThreads(m_ransacParallelNbThreads);
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~ ^~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/vision/src/key-point/vpKeyPoint.cpp:2641:36: warning: implicit conversion changes signedness: 'int' to 'unsigned long' [-Wsign-conversion]
 2641 |       char *path = new char[length + 1]; // char path[length + 1];
      |                    ~~~      ~~~~~~~^~~

+++ modules/vision/src/key-point/vpKeyPoint.cpp
@@ -809,7 +809,7 @@ bool vpKeyPoint::computePose(const std::vector<vpPoint> &objectVpPoints, vpHomog
 
   pose.setRansacFilterFlag(m_ransacFilterFlag);
   pose.setUseParallelRansac(m_ransacParallel);
-  pose.setNbParallelRansacThreads(m_ransacParallelNbThreads);
+  pose.setNbParallelRansacThreads(static_cast<int>(m_ransacParallelNbThreads));
   pose.setRansacNbInliersToReachConsensus(nbInlierToReachConsensus);
   pose.setRansacThreshold(m_ransacThreshold);
   pose.setRansacMaxTrials(m_nbRansacIterations);
@@ -2638,7 +2638,7 @@ void vpKeyPoint::loadLearningData(const std::string &filename, bool binaryMode, 
       int length = 0;
       vpIoTools::readBinaryValueLE(file, length);
       // Will contain the path to the training images
-      char *path = new char[length + 1]; // char path[length + 1];
+      char *path = new char[static_cast<size_t>(length) + 1]; // char path[length + 1];
 
       for (int cpt = 0; cpt < length; cpt++) {
         char c;
